数値の補正

+++ LanturnGame/Scene/GameMain.cpp
@@ -17,8 +17,6 @@ GameMain::GameMain()
 	for (int i = 0; i < GM_MAX_EFFECT_EXPLOSION; i++) {
 		explosion[i] = nullptr;
 	}
-	//explosion[0] = new Explosion;
-	//explosion[0]->SetLocation(bomb[0]->GetLocation());
 
 	background = new BackGround * [(int)pow((int)ceil(GM_MAX_MAPSIZE / 64.f) * 2, 2)];
 	for (int i = 0; i < (int)pow((int)ceil(GM_MAX_MAPSIZE / 64.f) * 2, 2); i++) {
@@ -190,7 +188,7 @@ AbstractScene* GameMain::Update()
 				if (i < MaxEnemyBomb) {
 					bomb[i] = new Bomb;
 					while (1) {
-						Vector2D spawnloc = (Vector2D((float)GetRand(MapSize * 2) - MapSize, (float)GetRand(MapSize * 2) - MapSize));
+						Vector2D spawnloc = (Vector2D((float)GetRand((int)MapSize * 2) - MapSize, (float)GetRand((int)MapSize * 2) - MapSize));
 						if (640 * (MapSize / GM_MAX_MAPSIZE) < fabsf(sqrtf(
 							powf((spawnloc.x - player->GetLocation().x), 2) +
 							powf((spawnloc.y - player->GetLocation().y), 2))))
@@ -239,7 +237,7 @@ AbstractScene* GameMain::Update()
 	if (MapSize > GM_MIN_MAPSIZE) {
 		MapSize -= MapCloseSpeed / 10;
 	}
-	MaxEnemyBomb = GM_MAX_ENEMY_BOMB * (MapSize / GM_MAX_MAPSIZE);
+	MaxEnemyBomb = (int)(GM_MAX_ENEMY_BOMB * (MapSize / GM_MAX_MAPSIZE));
 	game_frametime++;
 	CameraUpdate();
 	return this;
@@ -251,7 +249,7 @@ void GameMain::Draw() const
 
 	for (int i = 0; i < (int)pow((int)ceil(GM_MAX_MAPSIZE / 64.f) * 2, 2); i++) {
 		if (background[i] != nullptr) {
-			background[i]->Draw(player->GetLocation() + +Camerashake);
+			background[i]->Draw(player->GetLocation() + +(float)Camerashake);
 		}
 	}
 
@@ -269,7 +267,7 @@ void GameMain::Draw() const
 				powf((bomb[i]->GetLocation().x - player->GetLocation().x), 2) +
 				powf((bomb[i]->GetLocation().y - player->GetLocation().y), 2))))
 			{
-				bomb[i]->Draw(player->GetLocation() + +Camerashake);
+				bomb[i]->Draw(player->GetLocation() + +(float)Camerashake);
 			}
 		}
 	}
@@ -279,7 +277,7 @@ void GameMain::Draw() const
 				powf((explosion[i]->GetLocation().x - player->GetLocation().x), 2) +
 				powf((explosion[i]->GetLocation().y - player->GetLocation().y), 2))))
 			{
-				explosion[i]->Draw(player->GetLocation() + Camerashake);
+				explosion[i]->Draw(player->GetLocation() + (float)Camerashake);
 			}
 		}
 	}
@@ -338,7 +336,7 @@ void GameMain::SpawnExplosion(Vector2D loc) {
 
 void GameMain::CameraUpdate() {
 	if (CamerashakeCount > 0) {
-		Camerashake = round(CamerashakeCount / 2);
+		Camerashake = (int)round((double)CamerashakeCount / 2);
 		if (CamerashakeCount % 2 == 0) {
 			Camerashake *= -1;
 		}
@@ -351,5 +349,5 @@ void GameMain::SetCameraShake(int _i) {
 }
 
 void GameMain::SetMapSize(int i) {
-	MapSize = i;
+	MapSize = (float)i;
 }
